% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryHPD.R
\name{BinaryHPD}
\alias{BinaryHPD}
\title{Simulate HPD Width of Presence Data}
\usage{
BinaryHPD(n, p, mu.alpha = 0, sd.alpha = 1, mu.beta = 0,
  sd.beta = 1)
}
\arguments{
\item{n}{- number of sites.}

\item{p}{- number of species}

\item{mu.alpha}{- mean for alpha parameter: default is 0}

\item{sd.alpha}{- sd for alpha parameter: default is 1}

\item{mu.beta}{- mean for beta parameter: default is 0}

\item{sd.beta}{- sd for beta parameter: default is 1}
}
\value{
width - average HPD across all latent factors
}
\description{
function to simulate presence data and calculated HPD width for latent variables
}
