% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredibleViz.R
\name{CredibleViz}
\alias{CredibleViz}
\title{Visualize Uncertainty in Projection}
\usage{
CredibleViz(coord1, coord2, type = "points", items = NULL)
}
\arguments{
\item{coord1}{- mcmc object with first coordinate of projection}

\item{coord2}{- mcmc object with second coordinate of projection}

\item{type}{- plot type, options are 'points', 'circles', 'scatter', 'density}

\item{items}{- items to display uncertainty, only works for circles, density, or scatter}
}
\value{
plot.obj - ggplot object
}
\description{
This function simulates multivariate Poisson count data from a latent factor model.
}
\examples{
data(spider, package = 'mvabund')
spider.matrix <- matrix(as.numeric(spider$abund > 0), nrow = 28, ncol = 12)
ord <- ordinate_probit(500, spider.matrix)
CredibleViz(ord$z.samples[,,1], ord$z.samples[,,2], type = 'points', items = c(1,10,16))
CredibleViz(ord$z.samples[,,1], ord$z.samples[,,2], type = 'density', items = c(1,10,16))
}
