% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDwidth.R
\name{HPDwidth}
\alias{HPDwidth}
\title{HPD Width Simulation}
\usage{
HPDwidth(family, n, p, num.replicates = 1, mu.alpha = 0,
  sd.alpha = 1, mu.beta = 1, sd.beta = 1)
}
\arguments{
\item{family}{- family of the distribution, "binomial" and "poisson" are currently supported}

\item{n}{- number of sites.}

\item{p}{- number of species}

\item{num.replicates}{- number of simulations: default is 1}

\item{mu.alpha}{- mean for alpha parameter: default is 0}

\item{sd.alpha}{- sd for alpha parameter: default is 1}

\item{mu.beta}{- mean for beta parameter: default is 0}

\item{sd.beta}{- sd for beta parameter: default is 1}
}
\value{
width - average HPD across all latent factors
}
\description{
This function simulates multivariate data from a latent factor model to estimate the average width of the HPD intervals
}
