% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate_ordinal.R
\name{ordinate_ordinal}
\alias{ordinate_ordinal}
\title{Ordinal Ordination}
\usage{
ordinate_ordinal(num.mcmc, Y, burn.in = 100)
}
\arguments{
\item{num.mcmc}{- number of mcmc iterations}

\item{Y}{- an n X p matrix of ordinal data}

\item{burn.in}{- number of burn in samples, default is 100}
}
\value{
alpha.samples

beta.samples

theta.samples

z.samples
}
\description{
Fits latent factor model to ordinal data with cumulative probit sampling model
}
\examples{
set.seed(08052019)
Y <- matrix(base::sample(1:5, 50, replace = TRUE), 10, 5)
ordinate_ordinal(1000, Y)
}
