% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate_probit.R
\name{ordinate_probit}
\alias{ordinate_probit}
\title{Probit Ordination}
\usage{
ordinate_probit(num.mcmc, Y, burn.in = 100)
}
\arguments{
\item{num.mcmc}{- number of mcmc iterations}

\item{Y}{- an n X p matrix of presence data}

\item{burn.in}{- number of burn in samples, default is 100}
}
\value{
alpha.samples

beta.samples

theta.samples

z.samples
}
\description{
Fits latent factor model to binary data with probit link
}
