% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPoissonY.R
\name{simPoissonY}
\alias{simPoissonY}
\title{Count Data Simulation}
\usage{
simPoissonY(n, p, mu.alpha = 0, sd.alpha = 1, mu.beta = 0,
  sd.beta = 1)
}
\arguments{
\item{n}{- number of sites.}

\item{p}{- number of species}

\item{mu.alpha}{- mean for alpha parameter: default is 0}

\item{sd.alpha}{- sd for alpha parameter: default is 1}

\item{mu.beta}{- mean for beta parameter: default is 0}

\item{sd.beta}{- sd for beta parameter: default is 1}
}
\value{
mat - matrix representation of abundance

vec - vectorized representation of abundance

alpha.true - true values of alpha

beta.true - true values of alpha

theta.true - true values of alpha

z.true - true values of alpha
}
\description{
This function simulates multivariate Poisson count data from a latent factor model.
}
