% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simProbitY.R
\name{simProbitY}
\alias{simProbitY}
\title{Binary Data Simulation}
\usage{
simProbitY(n, p, mu.alpha = 0, sd.alpha = 1, mu.beta = 1, sd.beta = 1)
}
\arguments{
\item{n}{- number of sites.}

\item{p}{- number of species}

\item{mu.alpha}{- mean for alpha parameter: default is 0}

\item{sd.alpha}{- sd for alpha parameter: default is 1}

\item{mu.beta}{- mean for beta parameter: default is 0}

\item{sd.beta}{- sd for beta parameter: default is 1}
}
\value{
mat - matrix representation of presence / absence

vec - vectorized representation of presence / absence

alpha.true - true values of alpha

beta.true - true values of alpha

theta.true - true values of alpha

z.true - true values of alpha
}
\description{
This function simulates multivariate binary data from a latent factor model.
}
