% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredibleViz.R
\name{CredibleViz}
\alias{CredibleViz}
\title{Visualize Uncertainty in Projection}
\usage{
CredibleViz(coord1, coord2, type = "points", items = NULL)
}
\arguments{
\item{coord1}{- mcmc object with first coordinate of projection}

\item{coord2}{- mcmc object with second coordinate of projection}

\item{type}{- plot type, options are 'points', 'circles', 'scatter'}

\item{items}{- items to display uncertainty, only works for circles or scatter}
}
\value{
plot.obj - ggplot object
}
\description{
This function simulates multivariate Poisson count data from a latent factor model.
}
