% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate_poisson.R
\name{ordinate_poisson}
\alias{ordinate_poisson}
\title{Poisson Ordination}
\usage{
ordinate_poisson(num.mcmc, Y, burn.in = 100)
}
\arguments{
\item{num.mcmc}{- number of mcmc iterations}

\item{Y}{- an n X p matrix of abundance data}

\item{burn.in}{- number of burn in samples, default is 100}
}
\value{
alpha.samples

beta.samples

theta.samples

z.samples
}
\description{
Fits latent factor model to count data with Poisson sampling model
}
\examples{
set.seed(08052019)
Y <- rpois(100,100)
Y.matrix <- matrix(Y, 10, 10)
ordinate_poisson(1000, Y.matrix)
}
