% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate_probit.R
\name{ordinate_probit}
\alias{ordinate_probit}
\title{Probit Ordination}
\usage{
ordinate_probit(num.mcmc, Y.vec)
}
\arguments{
\item{num.mcmc}{- number of mcmc iterations}

\item{Y.vec}{- an n X p vectorized representation of presence data}
}
\value{
alpha.samples

beta.samples

theta.samples

z.samples
}
\description{
Fits latent factor model to binary data with probit link
}
